Fix docs around run (#131)
* docs: Correct usage of anon run function

The docs incorrectly suggested and showed usages of `run` with arity 3.
But `run` only allows for arity 1 and 2. This patch updates the docs to
show the correct usage and labeling for the function.

Fixes #128

* chore: Fix missing argument test

This test was warning when run. It unfortunately was getting a false
positive as written due to calling `.name` on a atom rather than getting
the `:name` property of the `:current_step` map.

This change fixes the test and looks for the correct error message.