fix(mermaid): use TB instead of TD
TD (top down) is an alias for TB (top to bottom) in `flowchart`, but the
alias is not present in subgraphs.

In reactors like the UserRegistration example, this results in broken
diagrams.

This change replaces TD with TB, allowing flowcharts and subgraphs to
function properly with a direction value of `:top_to_bottom`.